configs: prepare khadas-vim3{l}_ab_android for AOSP support
In AOSP, the bot VIM3 and VIM3L have 2 bootloader flavors,
depending on A/B enablement.

For example, for vim3l, the naming is:
- u-boot_kvim3l_noab.bin : legacy support
- u-boot_kvim3l_ab.bin   : A/B support

Prepare a defconfig to support u-boot_kvim3_ab.bin and u-boot_kvim3l_ab.bin.

This is identical to khadas-vim3{l}_ab_android but will be updated in the
next commit.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>